feat: load data from JSON file  and rename util functions
The dataset should really be external to the code. This keeps logical
changes separate from data-related changes and keeps the history a bit
more understandable.

Rename all of the functions to be internal (prefix with __) so the util
file is more consumable and obivous what is meant to be an internal and
external function.

Did some minor cleanup -- remove some silly semi-colons and renamed a
few variables so they are more obvious (e.g. gp -> games_played).

Lastly, note that some of the club abbreviations changed. Originaly I
used the abbreviations from /r/mls but now I'll use the mlssoccer.com
abbreviations.